Refactoy of splash screen
